+++ main.js
@@ -3,11 +3,13 @@ import App from './App'
 // #ifndef VUE3
 import Vue from 'vue'
 import './uni.promisify.adaptor'
+import uviewPlus, { setConfig } from 'uview-plus'
 Vue.config.productionTip = false
 App.mpType = 'app'
 const app = new Vue({
   ...App
 })
+app.use(uviewPlus)
 app.$mount()
 // #endif