+++ uni_modules/uview-plus/components/u-icon/u-icon.vue
@@ -72,7 +72,9 @@
 	export default {
 		name: 'u-icon',
 		beforeCreate() {
-			fontUtil.loadFont();
+			if (!fontUtil.params.loaded) {
+				fontUtil.loadFont();
+			}
     	},
 		data() {
 			return {
@@ -151,7 +153,6 @@
 </script>
 
 <style lang="scss" scoped>
-
 	// 变量定义
 	$u-icon-primary: $u-primary !default;
 	$u-icon-success: $u-success !default;
@@ -160,7 +161,7 @@
 	$u-icon-error: $u-error !default;
 	$u-icon-label-line-height:1 !default;
 
-	/* #ifdef MP-QQ || MP-TOUTIAO || MP-BAIDU || MP-KUAISHOU || MP-XHS */
+	/* #ifdef APP || MP-QQ || MP-TOUTIAO || MP-BAIDU || MP-KUAISHOU || MP-XHS */
 	// 2025/04/09在App/微信/支付宝/鸿蒙元服务已改用uni.loadFontFace加载字体
 	@font-face {
 		font-family: 'uicon-iconfont';