+++ uni_modules/uview-plus/components/u-table2/u-table2.vue
@@ -1,98 +1,159 @@
 <template>
-    <view scroll-x class="u-table2" :style="{ height: height ? height + 'px' : 'auto' }">
-        <!-- 表头 -->
-        <view v-if="showHeader" class="u-table-header" :class="{ 'u-table-sticky': fixedHeader }" :style="{minWidth: scrollWidth}">
-            <view class="u-table-row">
-                <view v-for="(col, colIndex) in columns" :key="col.key" class="u-table-cell"
-                    :style="headerColStyle(col)"
-					:class="[
-                        col.align ? 'u-text-' + col.align : '',
-                        headerCellClassName ? headerCellClassName(col) : '',
-                        col.fixed === 'left' ? 'u-table-fixed-left' : '',
-                        col.fixed === 'right' ? 'u-table-fixed-right' : ''
-                    ]" @click="handleHeaderClick(col)">
-					<slot name="header" :column="col" :columnIndex="colIndex" :level="1">
-					    {{ col.title }}
-					</slot>
-                    <view v-if="col.sortable">
-                        {{ getSortIcon(col.key) }}
+    <view class="u-table2" :class="{ 'u-table-border': border }">
+        <scroll-view scroll-x class="u-table2-content" :style="{ height: height ? height + 'px' : 'auto' }" @scroll="onScroll">
+            <!-- 表头 -->
+            <view v-if="showHeader" class="u-table-header" :class="{ 'u-table-sticky': fixedHeader }" :style="{minWidth: scrollWidth}">
+                <view class="u-table-row">
+                    <view v-for="(col, colIndex) in columns" :key="col.key" class="u-table-cell"
+                        :style="headerColStyle(col)"
+                        :class="[col.align ? 'u-text-' + col.align : '',
+                            headerCellClassName ? headerCellClassName(col) : '',
+                            getFixedClass(col)
+                        ]" @click="handleHeaderClick(col)">
+                        <slot name="header" :column="col" :columnIndex="colIndex" :level="1">
+                        </slot>
+                        <text v-if="!$slots['header']">{{ col.title }}</text>
+                        <template v-if="col.sortable">
+                            <slot name="headerSort" :sortStatus="getSortValue(col.key)" :column="col"
+                                :columnIndex="colIndex" :level="1">
+                            </slot>
+                            <view v-if="!$slots['headerSort']">
+                                {{ getSortIcon(col.key) }}
+                            </view>
+                        </template>
                     </view>
                 </view>
             </view>
-        </view>
 
-        <!-- 表体 -->
-        <view class="u-table-body" :style="{ minWidth: scrollWidth, maxHeight: maxHeight ? maxHeight + 'px' : 'none' }">
-            <template v-if="data && data.length > 0">
-                <template v-for="(row, index) in sortedData" :key="row[rowKey] || index">
-                    <view class="u-table-row" :class="[
-                        highlightCurrentRow && currentRow === row ? 'u-table-row-highlight' : '',
-                        rowClassName ? rowClassName(row, index) : '',
-                        stripe && index % 2 === 1 ? 'u-table-row-zebra' : ''
-                    ]" @click="handleRowClick(row)">
-                        <view v-for="(col, colIndex) in columns" :key="col.key"
-							class="u-table-cell" :class="[
-                            col.align ? 'u-text-' + col.align : '',
-                            cellClassName ? cellClassName(row, col) : '',
-                            col.fixed === 'left' ? 'u-table-fixed-left' : '',
-                            col.fixed === 'right' ? 'u-table-fixed-right' : ''
-                        ]" :style="cellStyleInner({row: row, column: col,
-							rowIndex: index, columnIndex: colIndex, level: 0})">
-                            <!-- 复选框列 -->
-                            <view v-if="col.type === 'selection'">
-                                <checkbox :checked="isSelected(row)"
-									@click.stop="toggleSelect(row)" />
-                            </view>
-
-                            <!-- 树形结构展开图标 -->
-                            <view v-else-if="col.type === 'expand'"
-								@click.stop="toggleExpand(row)">
-                                {{ isExpanded(row) ? '▼' : '▶' }}
-                            </view>
-
-                            <!-- 默认插槽或文本 -->
-                            <slot name="cell" :row="row" :column="col"
-								:rowIndex="index" :columnIndex="colIndex">
-                                <view class="u-table-cell_content">
-                                    {{ row[col.key] }}
-                                </view>
+            <!-- 表体 -->
+            <view class="u-table-body" :style="{ minWidth: scrollWidth, maxHeight: maxHeight ? maxHeight + 'px' : 'none' }">
+                <template v-if="data && data.length > 0">
+                    <table-row 
+                        v-for="(row, rowIndex) in sortedData"
+                        :key="row[rowKey] || rowIndex"
+                        :row="row" 
+                        :rowIndex="rowIndex"
+                        :parent-row="null"
+                        :columns="columns"
+                        :tree-props="treeProps"
+                        :row-key="rowKey"
+                        :expanded-keys="expandedKeys"
+                        :cell-style-inner="cellStyleInner"
+                        :is-expanded="isExpanded"
+                        :row-class-name="rowClassName"
+                        :stripe="stripe"
+                        :cell-class-name="cellClassName"
+                        :get-fixed-class="getFixedClass"
+                        :highlight-current-row="highlightCurrentRow"
+                        :current-row="currentRow"
+                        :handle-row-click="handleRowClick"
+                        :toggle-expand="toggleExpand"
+                        :level="1"
+                        :rowHeight="rowHeight"
+                        :hasTree="hasTree"
+                        :selectedRows="selectedRows"
+                        :expandWidth="expandWidth"
+                        :computedMainCol="computedMainCol"
+                        @toggle-select="toggleSelect"
+                        @row-click="handleRowClick"
+                        @toggle-expand="toggleExpand"
+                    >
+                        <template v-slot:cellChild="scope">
+                            <slot name="cell" :row="scope.row" :column="scope.column" :prow="scope.prow"
+                                :rowIndex="scope.rowIndex" :columnIndex="scope.columnIndex" :level="scope.level">
+                            </slot>                      
+                        </template>
+                    </table-row>
+                </template>
+                <template v-else>
+                    <slot name="empty">
+                    </slot>
+                    <view v-if="!$slots['empty']" class="u-table-empty">{{ emptyText }}</view>
+                </template>
+            </view>
+        </scroll-view>
+        
+        <!-- 固定列浮动视图 -->
+        <view v-if="showFixedColumnShadow" class="u-table-fixed-shadow" :style="{ height: tableHeight }">
+            <!-- 表头 -->
+            <view v-if="showHeader" class="u-table-header" :class="{ 'u-table-sticky': fixedHeader }" :style="{minWidth: scrollWidth}">
+                <view class="u-table-row" :style="{height: headerHeight}">
+                    <view v-for="(col, colIndex) in visibleFixedLeftColumns" :key="col.key" class="u-table-cell"
+                        :style="headerColStyle(col)"
+                        :class="[col.align ? 'u-text-' + col.align : '',
+                            headerCellClassName ? headerCellClassName(col) : '',
+                            getFixedClass(col)
+                        ]" @click="handleHeaderClick(col)">
+                        <slot name="header" :column="col" :columnIndex="colIndex" :level="1">
+                        </slot>
+                        <text v-if="!$slots['header']">{{ col.title }}</text>
+                        <template v-if="col.sortable">
+                            <slot name="headerSort" :sortStatus="getSortValue(col.key)" :column="col"
+                                :columnIndex="colIndex" :level="1">
                             </slot>
-                        </view>
-                    </view>
-
-                    <!-- 子级渲染 -->
-                    <template v-if="isExpanded(row) && row[treeProps.children] && row[treeProps.children].length">
-                        <view v-for="childRow in row[treeProps.children]" :key="childRow[rowKey]"
-                            class="u-table-row u-table-row-child">
-                            <view v-for="(col2, col2Index) in columns" :key="col2.key" class="u-table-cell"
-                                :style="cellStyleInner({row: childRow, column: col2,
-									rowIndex: index, columnIndex: col2Index, level: 1})">
-                                <slot name="cell" :row="childRow" :column="col2" :prow="row"
-									:rowIndex="index" :columnIndex="col2Index" :level="1">
-                                    <view class="u-table-cell_content">
-                                        {{ childRow[col2.key] }}
-                                    </view>
-                                </slot>
+                            <view v-if="!$slots['headerSort']">
+                                {{ getSortIcon(col.key) }}
                             </view>
-                        </view>
+                        </template>
+                    </view>
+                </view>
+            </view>
+
+            <!-- 表体 -->
+            <view class="u-table-body" :style="{ minWidth: scrollWidth, maxHeight: maxHeight ? maxHeight + 'px' : 'none' }">
+                <template v-if="data && data.length > 0">
+                    <template v-for="(row, rowIndex) in sortedData" :key="row[rowKey] || rowIndex">
+                        <!-- 子级渲染 (递归组件) -->
+                        <table-row 
+                            :row="row" 
+                            :rowIndex="rowIndex"
+                            :parent-row="null"
+                            :columns="visibleFixedLeftColumns"
+                            :tree-props="treeProps"
+                            :row-key="rowKey"
+                            :expanded-keys="expandedKeys"
+                            :cell-style-inner="cellStyleInner"
+                            :is-expanded="isExpanded"
+                            :row-class-name="rowClassName"
+                            :stripe="stripe"
+                            :cell-class-name="cellClassName"
+                            :get-fixed-class="getFixedClass"
+                            :highlight-current-row="highlightCurrentRow"
+                            :current-row="currentRow"
+                            :handle-row-click="handleRowClick"
+                            :toggle-expand="toggleExpand"
+                            :level="1"
+                            :rowHeight="rowHeight"
+                            :hasTree="hasTree"
+                            :selectedRows="selectedRows"
+                            :expandWidth="expandWidth"
+                            :computedMainCol="computedMainCol"
+                            @toggle-select="toggleSelect"
+                            @row-click="handleRowClick"
+                            @toggle-expand="toggleExpand"
+                        >
+                            <template v-slot:cellChild="scope">
+                                <slot name="cell" :row="scope.row" :column="scope.column" :prow="scope.prow"
+                                    :rowIndex="scope.rowIndex" :columnIndex="scope.columnIndex" :level="scope.level">
+                                </slot>                      
+                            </template>
+                        </table-row>
                     </template>
                 </template>
-            </template>
-            <template v-else>
-                <slot name="empty">
-                    <view class="u-table-empty">{{ emptyText }}</view>
-                </slot>
-            </template>
+            </view>
         </view>
     </view>
 </template>
 
 <script>
-import { ref, watch, computed } from 'vue'
 import { addUnit, sleep } from '../../libs/function/index';
+import tableRow from './tableRow.vue'; // 引入递归组件
 
 export default {
     name: 'u-table2',
+    components: {
+        tableRow // 注册递归组件
+    },
     props: {
         data: {
             type: Array,
@@ -227,6 +288,19 @@ export default {
             type: String,
             default: '暂无数据'
         },
+        // 添加mainCol属性，用于指定树形结构展开控制图标所在的列
+        mainCol: {
+            type: String,
+            default: ''
+        },
+        expandWidth: {
+            type: String,
+            default: '25px'
+        },
+        rowHeight: {
+            String,
+            default: '40px'
+        }
     },
     emits: [
         'select', 'select-all', 'selection-change',
@@ -236,35 +310,179 @@ export default {
     ],
     data() {
         return {
-            scrollWidth: 'auto'
+            scrollWidth: 'auto',
+            // 将setup中的ref转换为data属性
+            expandedKeys: [...this.expandRowKeys],
+            selectedRows: [],
+            sortConditions: [],
+            currentRow: null,
+            scrollLeft: 0, // 新增滚动位置数据
+            showFixedColumnShadow: false, // 是否显示固定列阴影
+            fixedLeftColumns: [], // 左侧固定列
+            tableHeight: 'auto', // 表格高度
+            headerHeight: 'auto', // 新增表头高度属性
+            hasTree: false // 新增属性，用于判断是否存在树形结构
         }
     },
     mounted() {
         this.getComponentWidth()
+        // 处理currentRowKey初始化
+        if (this.currentRowKey !== null) {
+            const found = this.data.find(item => item[this.rowKey] === this.currentRowKey);
+            if (found) {
+                this.currentRow = found;
+            }
+        }
+        // 获取固定列
+        this.fixedLeftColumns = this.columns.filter(col => col.fixed === 'left');
     },
-	computed: {
-	},
+    computed: {
+        // 将setup中的computed转换为computed属性
+        filteredData() {
+            return this.data.filter(row => {
+                return Object.keys(this.filters).every(key => {
+                    const filter = this.filters[key];
+                    if (!filter) return true;
+                    return row[key]?.toString().includes(filter.toString());
+                });
+            });
+        },
+        sortedData() {
+            if (!this.sortConditions.length) return this.filteredData;
+
+            const data = [...this.filteredData];
+
+            return data.sort((a, b) => {
+                for (const condition of this.sortConditions) {
+                    const { field, order } = condition;
+                    let valA = a[field];
+                    let valB = b[field];
+
+                    if (this.sortMethod) {
+                        const result = this.sortMethod(a, b, field);
+                        if (result !== 0) return result * (order === 'ascending' ? 1 : -1);
+                    }
+
+                    if (valA < valB) return order === 'ascending' ? -1 : 1;
+                    if (valA > valB) return order === 'ascending' ? 1 : -1;
+                }
+                return 0;
+            });
+        },
+        // 计算当前应该显示的固定左侧列
+        visibleFixedLeftColumns() {
+            if (this.scrollLeft <= 0) {
+                return [];
+            }
+            
+            let totalWidth = 0;
+            let fixedWidth = 0;
+            const visibleColumns = [];
+            
+            // 遍历所有列，不仅仅是固定列
+            for (let i = 0; i < this.columns.length; i++) {
+                const col = this.columns[i];
+                const colWidth = col.width ? parseInt(col.width) : 100; // 默认宽度100px
+                
+                // 如果是固定列且滚动位置足够显示该列
+                if (col.fixed === 'left' && this.scrollLeft > totalWidth - fixedWidth) {
+                    visibleColumns.push(col);
+                    fixedWidth += colWidth;
+                }
+                
+                totalWidth += colWidth;
+            }
+            
+            return visibleColumns;
+        },
+        // 获取mainCol的值，如果未设置则默认为第一列的key
+        computedMainCol() {
+            if (this.mainCol) {
+                return this.mainCol;
+            }
+            // 修改为排除有type值的列
+            const validColumns = this.columns.filter(col => !col.type);
+            let mainCol = validColumns && validColumns.length > 0 ? validColumns[0].key : '';
+            // console.log('mainCol', mainCol)
+            return mainCol;
+        }
+    },
+    watch: {
+        // 将setup中的watch转换为watch属性
+        expandRowKeys: {
+            handler(newVal) {
+                this.expandedKeys = [...newVal];
+            },
+            immediate: true
+        },
+        currentRowKey: {
+            handler(newVal) {
+                const found = this.data.find(item => item[this.rowKey] === newVal);
+                if (found) {
+                    this.currentRow = found;
+                }
+            },
+            immediate: true
+        },
+        columns: {
+            handler() {
+                // this.fixedLeftColumns = this.columns.filter(col => col.fixed === 'left');
+            },
+            deep: true,
+            immediate: false
+        }
+    },
     methods: {
         addUnit,
-		headerColStyle(col) {
-			let style = {
-				width: col.width ? addUnit(col.width) : 'auto',
-				flex: col.width ? 'none' : 1
-			};
-			if (col?.style) {
-				style = {...style, ...col?.style};
-			}
-			return style;
-		},
+        onScroll(e) {
+            this.scrollLeft = e.detail.scrollLeft;
+            // 获取所有左侧固定列
+            this.fixedLeftColumns = this.columns.filter(col => col.fixed === 'left');
+            // 计算是否需要显示固定列阴影
+            if (this.fixedLeftColumns.length > 0) {
+                this.showFixedColumnShadow = this.scrollLeft > 0;
+            }
+        },
+        
+        getFixedShadowStyle(col, index) {
+            let style = {
+                width: col.width ? addUnit(col.width) : 'auto',
+            };
+            
+            if (col?.style) {
+                style = {...style, ...col?.style};
+            }
+            
+            return style;
+        },
+        
+        getFixedClass(col) {
+            return ''; // 不再使用原来的固定列样式类
+        },
+        
+        headerColStyle(col) {
+            let style = {
+                width: col.width ? addUnit(col.width) : 'auto',
+                flex: col.width ? 'none' : 1
+            };
+            if (col?.style) {
+                style = {...style, ...col?.style};
+            }
+            return style;
+        },
+        
 		setCellStyle(e) {
 			this.cellStyle = e
 		},
 		cellStyleInner(scope) {
 			let style = {
 				width: scope.column?.width ? addUnit(scope.column.width) : 'auto',
-				flex: scope.column?.width ? 'none' : 1,
-				paddingLeft: (24 * scope.level) + 'px'
+				flex: scope.column?.width ? 'none' : 1
 			};
+            // 只有展开列设置padding
+            if (scope.column.key == this.computedMainCol) {
+                style.paddingLeft = (16 * (scope.level -1 )) + 2 + 'px'
+            }
 			if (this.cellStyle != null) {
 				let styleCalc = this.cellStyle(scope)
 				if (styleCalc != null) {
@@ -280,156 +498,129 @@ export default {
 			this.$uGetRect('.u-table-row').then(size => {
 				this.scrollWidth = size.width + 'px'
 			})
+            
+            // 获取表头高度并设置
+            this.$uGetRect('.u-table-header').then(size => {
+                if (size.height) {
+                    this.headerHeight = size.height + 'px';
+                }
+            })
+            
+            // 遍历数据列表第一层判断是否存在树形结构
+            this.hasTree = this.sortedData.some(item => {
+                return item[this.treeProps.children] && item[this.treeProps.children].length > 0;
+            });
 		},
-    },
-    setup(props, { emit }) {
-        const expandedKeys = ref([...props.expandRowKeys]);
-        const selectedRows = ref([]);
-        const sortConditions = ref([]);
-
-        // 当前高亮行
-        const currentRow = ref(null);
-
-        watch(
-            () => props.expandRowKeys,
-            newVal => {
-                expandedKeys.value = [...newVal];
+        // 将setup中的函数转换为methods
+        handleRowClick(row) {
+            if (this.highlightCurrentRow) {
+                const oldRow = this.currentRow;
+                this.currentRow = row;
+                this.$emit('current-change', row, oldRow);
             }
-        );
-
-        watch(
-            () => props.currentRowKey,
-            newVal => {
-                const found = props.data.find(item => item[props.rowKey] === newVal);
-                if (found) {
-                    currentRow.value = found;
-                }
-            }
-        );
-
-        // 过滤后的数据
-        const filteredData = computed(() => {
-            return props.data.filter(row => {
-                return Object.keys(props.filters).every(key => {
-                    const filter = props.filters[key];
-                    if (!filter) return true;
-                    return row[key]?.toString().includes(filter.toString());
-                });
-            });
-        });
-
-        // 排序后的数据
-        const sortedData = computed(() => {
-            if (!sortConditions.value.length) return filteredData.value;
-
-            const data = [...filteredData.value];
-
-            return data.sort((a, b) => {
-                for (const condition of sortConditions.value) {
-                    const { field, order } = condition;
-                    let valA = a[field];
-                    let valB = b[field];
-
-                    if (props.sortMethod) {
-                        const result = props.sortMethod(a, b, field);
-                        if (result !== 0) return result * (order === 'ascending' ? 1 : -1);
-                    }
-
-                    if (valA < valB) return order === 'ascending' ? -1 : 1;
-                    if (valA > valB) return order === 'ascending' ? 1 : -1;
-                }
-                return 0;
-            });
-        });
-
-        function handleRowClick(row) {
-            if (props.highlightCurrentRow) {
-                const oldRow = currentRow.value;
-                currentRow.value = row;
-                emit('current-change', row, oldRow);
-            }
-            emit('row-click', row);
-        }
-
-        function handleHeaderClick(column) {
+            this.$emit('row-click', row);
+        },
+        handleHeaderClick(column) {
             if (!column.sortable) return;
 
-            const index = sortConditions.value.findIndex(c => c.field === column.key);
+            const index = this.sortConditions.findIndex(c => c.field === column.key);
             let newOrder = 'ascending';
 
             if (index >= 0) {
-                if (sortConditions.value[index].order === 'ascending') {
+                if (this.sortConditions[index].order === 'ascending') {
                     newOrder = 'descending';
                 } else {
-                    sortConditions.value.splice(index, 1);
-                    emit('sort-change', sortConditions.value);
+                    this.sortConditions.splice(index, 1);
+                    this.$emit('sort-change', this.sortConditions);
                     return;
                 }
             }
 
-            if (!props.multiSort) {
-                sortConditions.value = [{ field: column.key, order: newOrder }];
+            if (!this.multiSort) {
+                this.sortConditions = [{ field: column.key, order: newOrder }];
             } else {
                 if (index >= 0) {
-                    sortConditions.value[index].order = newOrder;
+                    this.sortConditions[index].order = newOrder;
                 } else {
-                    sortConditions.value.push({ field: column.key, order: newOrder });
+                    this.sortConditions.push({ field: column.key, order: newOrder });
                 }
             }
 
-            emit('sort-change', sortConditions.value);
-        }
-
-        function getSortIcon(field) {
-            const cond = sortConditions.value.find(c => c.field === field);
+            this.$emit('sort-change', this.sortConditions);
+        },
+        getSortIcon(field) {
+            const cond = this.sortConditions.find(c => c.field === field);
             if (!cond) return '';
             return cond.order === 'ascending' ? '↑' : '↓';
-        }
-
-        function toggleSelect(row) {
-            const index = selectedRows.value.findIndex(r => r[props.rowKey] === row[props.rowKey]);
+        },
+        getSortValue(field) {
+            const cond = this.sortConditions.find(c => c.field === field);
+            if (!cond) return '';
+            return cond.order === 'ascending';
+        },
+        toggleSelect(row) {
+            const index = this.selectedRows.findIndex(r => r[this.rowKey] === row[this.rowKey]);
             if (index >= 0) {
-                selectedRows.value.splice(index, 1);
+                // 取消选中当前行及其所有子节点
+                this.selectedRows.splice(index, 1);
+                // 递归取消所有子节点
+                this.unselectChildren(row);
             } else {
-                selectedRows.value.push(row);
+                // 选中当前行及其所有子节点
+                this.selectedRows.push(row);
+                // 递归选中所有子节点
+                this.selectChildren(row);
             }
-            emit('selection-change', selectedRows.value);
-            emit('select', row);
-        }
-
-        function isSelected(row) {
-            return selectedRows.value.some(r => r[props.rowKey] === row[props.rowKey]);
-        }
-
-        function toggleExpand(row) {
-            const key = row[props.rowKey];
-            const index = expandedKeys.value.indexOf(key);
+            console.log(this.selectedRows)
+            this.$emit('selection-change', this.selectedRows);
+            this.$emit('select', row);
+        },
+        toggleExpand(row) {
+            // console.log(row)
+            const key = row[this.rowKey];
+            const index = this.expandedKeys.indexOf(key);
             if (index === -1) {
-                expandedKeys.value.push(key);
+                this.expandedKeys.push(key);
             } else {
-                expandedKeys.value.splice(index, 1);
+                this.expandedKeys.splice(index, 1);
             }
-            emit('expand-change', expandedKeys.value);
-        }
-
-        function isExpanded(row) {
-            return expandedKeys.value.includes(row[props.rowKey]);
-        }
-
-        return {
-            currentRow,
-            sortedData,
-            expandedKeys,
-            selectedRows,
-            sortConditions,
-            handleRowClick,
-            handleHeaderClick,
-            getSortIcon,
-            toggleSelect,
-            isSelected,
-            toggleExpand,
-            isExpanded
-        };
+            this.$emit('expand-change', this.expandedKeys);
+        },
+        isExpanded(row) {
+            if (!row) {
+                return false;
+            }
+            return this.expandedKeys.includes(row[this.rowKey]);
+        },
+        // 新增方法：递归选中所有子节点
+        selectChildren(row) {
+            const children = row[this.treeProps.children];
+            if (children && children.length > 0) {
+                children.forEach(child => {
+                    // 检查是否已选中，避免重复添加
+                    const childIndex = this.selectedRows.findIndex(r => r[this.rowKey] === child[this.rowKey]);
+                    if (childIndex === -1) {
+                        this.selectedRows.push(child);
+                    }
+                    // 递归处理子节点的子节点
+                    this.selectChildren(child);
+                });
+            }
+        },
+        // 新增方法：递归取消选中所有子节点
+        unselectChildren(row) {
+            const children = row[this.treeProps.children];
+            if (children && children.length > 0) {
+                children.forEach(child => {
+                    const childIndex = this.selectedRows.findIndex(r => r[this.rowKey] === child[this.rowKey]);
+                    if (childIndex >= 0) {
+                        this.selectedRows.splice(childIndex, 1);
+                    }
+                    // 递归处理子节点的子节点
+                    this.unselectChildren(child);
+                });
+            }
+        },
     }
 };
 </script>
@@ -439,6 +630,7 @@ export default {
     width: auto;
     overflow: auto;
     white-space: nowrap;
+    position: relative;
 
     .u-table-header {
         min-width: 100% !important;
@@ -449,6 +641,7 @@ export default {
     .u-table-body {
         min-width: 100% !important;
         width: fit-content;
+        position: relative;
     }
 
     .u-table-sticky {
@@ -461,31 +654,36 @@ export default {
         display: flex;
         flex-direction: row;
         align-items: center;
-        border-bottom: 1rpx solid #ebeef5;
+        border-bottom: 1px solid #ebeef5;
         overflow: hidden;
+        position: relative;
+        // min-height: 40px;
+    }
+
+    // 添加border样式支持
+    &.u-table-border {
+        border-top: 1px solid #ebeef5;
+        border-left: 1px solid #ebeef5;
+        border-right: 1px solid #ebeef5;
+        .u-table-cell {
+            border-right: 1px solid #ebeef5;
+        }
+        
+        .u-table-cell:last-child {
+            border-right: none;
+        }
     }
 
     .u-table-cell {
         flex: 1;
         display: flex;
         flex-direction: row;
-        padding: 5px 4px;
+        padding: 10px 10px;
         font-size: 14px;
         white-space: nowrap;
         overflow: hidden;
         text-overflow: ellipsis;
-    }
-
-    .u-table-fixed-left {
-        position: sticky;
-        left: 0;
-        z-index: 9;
-    }
-
-    .u-table-fixed-right {
-        position: sticky;
-        right: 0;
-        z-index: 9;
+        line-height: 1.1;
     }
 
     .u-table-row-zebra {
@@ -512,6 +710,37 @@ export default {
 
     .u-text-right {
         text-align: right;
+    }
+}
+
+// 固定列浮动视图
+.u-table-fixed-shadow {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: auto;
+    z-index: 20;
+    box-shadow: 2px 0 5px rgba(0, 0, 0, 0.15);
+    overflow: hidden;
+    background-color: #ffffff;
+}
+
+// .u-table-fixed-row {
+//     display: flex;
+//     flex-direction: row;
+//     align-items: center;
+//     border-bottom: 1rpx solid #ebeef5;
+//     position: relative;
+// }
+
+// 为固定列也添加border样式支持
+.u-table-fixed-shadow .u-table-border {
+    .u-table-cell {
+        border-right: 1rpx solid #ebeef5;
+    }
+    
+    .u-table-cell:last-child {
+        border-right: none;
     }
 }
 </style>