+++ uni_modules/uview-plus/libs/function/index.js
@@ -66,10 +66,10 @@ export function sys() {
 }
 export function getWindowInfo() {
 	let ret = {}
-	// #ifdef APP || H5 || MP-WEIXIN
+	// #ifdef APP || H5
 	ret = uni.getWindowInfo()
 	// #endif
-	// #ifndef APP || H5 || MP-WEIXIN
+	// #ifndef APP || H5
 	ret = sys()
 	// #endif
 	return ret
@@ -360,6 +360,10 @@ export function timeFormat(dateTime = null, formatStr = 'yyyy-mm-dd') {
   // 若用户传入字符串格式时间戳，new Date无法解析，需做兼容
   else if (typeof dateTime === 'string' && /^\d+$/.test(dateTime.trim())) {
     date = new Date(Number(dateTime))
+  }
+  // 检查是否为UTC格式的时间字符串 (2024-12-18T02:25:31.432Z)
+  else if (typeof dateTime === 'string' && /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$/.test(dateTime)) {
+    date = new Date(dateTime)
   }
   // 其他都认为符合 RFC 2822 规范
   else {